new elevator (#4)
## Description

Please include a summary of the changes in this PR. If it is fixing an
issue, describe the issue. Please also include relevant motivation and
context such that a newer programmer could understand your changes just
by reading this description.

PID elevator attached to the coral intake. Elevator can go up and down. To resolve conflicts, code from main was pulled and incoming changes were accepted.


## How Has This Been Tested?

Please describe the tests that you ran to verify your changes. Be
detailed so that your code reviewer can understand exactly how much and
what kinds of testing were done, and which might still be worthwhile to
do.

Untested (for now)

- [ ] Unit tests: [Add your description here]
- [ ] Simulator testing: [Add your description here]
- [ ] On-robot bench testing: [Add your description here]
- [ ] On-robot field testing: [Add your description here]

---------

Co-authored-by: xavier <avier@gmail.com>
Co-authored-by: DS <team766robotics@gmail.com>
Co-authored-by: Max Spier <68516760+maxspier@users.noreply.github.com>