Remove device object cache so there aren't CAN id conflicts (#34)
The electrical team wants to assign unique CAN device IDs to each motor
in our inventory. Currently, our software framework assumes that all
motors are in the same ID pool, but this is incorrect: each device
manufacturer gets a separate pool of IDs, each of which supports up to
64 devices; there are additional manufacturer ID bits (hidden from the
user) that are added to create the complete device ID. We now have more
than 64 total motors in the shop, so this PR removes the software
restriction that all motors must have unique device IDs so that we can
have up to 64 motors from each manufacturer.