AspectJ for automatic checkContextReservation and log mechanism methods (#37)
`MechanismsAspect` ensures that all public methods of a `Mechanism`
class can only be called by `Rule`s or `Procedure`s that have reserved
that `Mechanism`. It also automatically logs all calls to any method
that it checks, including the value of any arguments that are passed to
the method. The intent of this logging is that we have a record of all
operations that impose a change in state of the `Mechanism`, which
complements automatic logging of `Mechanism` `Status`es which logs the
natural evolution of the `Mechanism`'s state.

If a public method may be called without reserving the `Mechanism`,
apply the `NoReservationRequired` annotation to that method, but you
probably won't need to do that.

This uses AspectJ to insert some additional code at the beginning of
each public method of every `Mechanism` class while the class is being
compiled. Google for AspectJ or "aspect-oriented programming" for more
information.