Convert TimestampedApriltag to a record type for better logging (#35)
AdvantageKit supports automatically logging [`record`
types](https://www.baeldung.com/java-record-keyword), as well as a
number of types builtin to Wpilib, such as `Pose3d` (but `AprilTag` is
not one of those). This PR converts `KalamanApriltag` to a record type
to allow it to be automatically logged (as well as sent to Dashboards).
Also renames it to `VisionApriltag` to better describe its use as a
value published by Vision.