add whenStatusMatching convenience helper (#40)
## Description

simple helpers to return `BooleanSupplier`s for checking that a Status
matches a predicate / a recent Status matches the predicate. useful for
Rule predicates.

## How Has This Been Tested?

Please describe the tests that you ran to verify your changes. Be
detailed so that your code reviewer can understand exactly how much and
what kinds of testing were done, and which might still be worthwhile to
do.

- [x] Unit tests: added test for `whenStatusMatching`
- [ ] Simulator testing: [Add your description here]
- [ ] On-robot bench testing: [Add your description here]
- [ ] On-robot field testing: [Add your description here]

---------

Co-authored-by: Ryan Cahoon <ryan.m.cahoon@gmail.com>