added mechanisms with status to coralintake and elevator (#39)
## Description

Please include a summary of the changes in this PR. If it is fixing an
issue, describe the issue. Please also include relevant motivation and
context such that a newer programmer could understand your changes just
by reading this description.

Added MechanismsWithStatus to coral intake and elevator java. Updated to
framework 3 aswell.

## How Has This Been Tested?

Not yet tested.

Please describe the tests that you ran to verify your changes. Be
detailed so that your code reviewer can understand exactly how much and
what kinds of testing were done, and which might still be worthwhile to
do.

- [ ] Unit tests: [Add your description here]
- [ ] Simulator testing: [Add your description here]
- [ ] On-robot bench testing: [Add your description here]
- [ ] On-robot field testing: [Add your description here]

---------

Co-authored-by: xavier <avier@gmail.com>