SparkMax: Better handling of sensor inversion (#86)
The Rev API fails if we try to set the inversion of the encoder built
into the brushless motors. This is because the encoders are
factory-assembled into the motors in the correct orientation, and it
makes no sense to try to set the encoder to the opposite direction of
the motor's movement.

We had previously
(https://github.com/Team766/2025/commit/ed7a770e6358cbe650a49d3cac86d0de32051450)
hacked around this by only setting the sensor inversion on the
SparkMaxes when invert=true. However, this has the opportunity for
confusion if someone has set inversion on a sensor that supports it
(e.g. an external encoder) and is then trying to set the sensor back to
un-inverted (e.g. by changing the robot config file). In this case, we
do want to set inverted=false with the Rev API to undo the previous set
inverted=true.

This PR improves the handling of this situation. Instead of skipping all
calls to `setSensorInverted(false)`, we instead skip all calls which try
to set the inversion of specifically the built-in encoder. However, we
do still sanity-check that the user only tries to set the inversion of
the built-in encoder to false, reporting an error if trying to invert
the built-in encoder.