Fix exception during init from LEDString (#107)
https://github.com/Team766/2025/pull/32 migrated `LEDString` to be a
`MultiFacetedMechanism`. This allows segments of the LED string to be
controlled separately. When setting up the `LEDString`, you call its
`makeSegment` method to create these segments. During initialization,
the code does not hold a reservation on the mechanism, so we need to
annotate `makeSegment` as `@NoReservationRequired` to skip the mechanism
reservation checks.

This resolves an initialization failure that was seen on Bearracuda.

`makeSegment` should be safe to skip reservation checks: reasonable code
should not call it after initialization; but if any code does, it has
checks to make sure that the new segment doesn't overlap any existing
segments, so it can't be used to take control of LEDs that are already
reserved.