Add exception safety around RobotConfigurator.getAutonomousModes (#106)
## Description

@maxspier recently encountered some sort of failure in
`GenericRobotMain`'s constructor. I didn't get access to the stacktrace
from the exception, so I wrote some unit tests that try to create
pathological cases. The only thing I could find is if there's an
exception in `getAutonomousModes` or it returns `null`, that could cause
`GenericRobotMain`'s constructor to fail, so this PR adds some
protection against that. Not sure if this will solve Max's issue, but it
seems a good improvement to make anyway.

## How Has This Been Tested?

Please describe the tests that you ran to verify your changes. Be
detailed so that your code reviewer can understand exactly how much and
what kinds of testing were done, and which might still be worthwhile to
do.

- [X] Unit tests: [Add your description here]
- [ ] Simulator testing: [Add your description here]
- [ ] On-robot bench testing: [Add your description here]
- [ ] On-robot field testing: [Add your description here]

Co-authored-by: Max Spier <68516760+maxspier@users.noreply.github.com>