Get the 3d simulator working with MF3 (#112)
Our code base is evolving to use more WPILib features directly, rather
than treating WPILib as a hardware interface implementation detail. This
means we need to setup a proper WPILib environment for our code to run
in, even when our code is running in simulations instead of on real
robots. Luckily, WPILib does have reasonable support for simulation that
we can integrate with (launched via `./gradlew simulateJava`). However,
WPILib wants to take control of overall execution of the program once
the robot code starts running, so we need to refactor our simulation
code to run as a "periodic" function within WPILib - the robot code is
now calling the simulation code, rather than the simulation code calling
the robot code.

This PR contains enough changes to get MF3 code running with the 3d
simulator, but is not a full conversion - the simulator is still
basically ignoring WPILib in most aspects. Most significantly, we're
still swapping to a different implementation of RobotProvider and its
motor and sensor classes in simulation. Work should be done to convert
to using the WPILib simulation device interface so that our code can
access vendor device APIs directly if we want to.