Mark framework files as readonly to prevent accidental changes (#122)
## Description

A significant number of the issues that I got pulled in to fix in rookie
code this year were due to incorrect applications of VSCode's
auto-correct features (e.g. create a missing method in a class) causing
changes to framework code. I know other veterans and mentors had to deal
with these issues as well. Perhaps the ultimate answer is to move the
framework out of the main repo, but unless/until then, this provides a
pretty reasonable stopgap.

This change marks code in framework directories as readonly in VSCode.
Attempting to directly modify a readonly file will present a popup
tooltip stating that the file is readonly. An easy option to override
the restriction is made available, if the user feels confident they know
what they're doing (so changes to the framework by specialists and
mentors are still very doable):
<img width="617" height="78" alt="image"
src="https://github.com/user-attachments/assets/fb264f88-778f-4ac1-8cf9-a2a140b608b9"
/>
Making the files readonly also seems to prevent the auto-correct
operations from making any changes, as intended (though unfortunately
the little lightbulb still presents them as available options to the
user).

## How Has This Been Tested?

Tried editing framework files and confirmed that the tooltip is
displayed but no changes are allowed. After clicking "Click here to set
writeable" I am able to make changes.

Confirmed that when the files are readonly, the quick fix (lightbulb)
options don't actually make changes to the files when selected. One
confusing thing is that if the readonly file is currently open in an
editor tab, the quick fix _does_ appear to make changes to the file
which makes the little red squiggle go away. However, the changes are
not actually written to the file on-disk, so a compiler error will still
be reported when the code is built (i.e. with gradle). Closing the
editor tab will discard the (in-memory only) changes and make the red
squiggle come back. Filed
https://github.com/redhat-developer/vscode-java/issues/4275 for this UX
issue.